Move legacy container finding into utils (for now) and update tests
