+++ nf_core/pipelines/download/utils.py
@@ -4,8 +4,11 @@ import enum
 import io
 import logging
 import os
+import re
 import tempfile
-from typing import Callable, Dict, Generator, Iterable, List, Optional, Tuple
+import textwrap
+from collections.abc import Generator, Iterable
+from typing import Callable, Optional
 
 import requests
 import requests_cache
@@ -147,16 +150,16 @@ class FileDownloader:
 
     def download_files_in_parallel(
         self,
-        download_files: Iterable[Tuple[str, str]],
+        download_files: Iterable[tuple[str, str]],
         parallel_downloads: int,
-        callback: Optional[Callable[[Tuple[str, str], Status], None]] = None,
-    ) -> List[Tuple[str, str]]:
+        callback: Optional[Callable[[tuple[str, str], Status], None]] = None,
+    ) -> list[tuple[str, str]]:
         """Download multiple files in parallel.
 
         Args:
-            download_files (Iterable[Tuple[str, str]]): List of tuples with the remote URL and the local output path.
+            download_files (Iterable[tuple[str, str]]): list of tuples with the remote URL and the local output path.
             parallel_downloads (int): Number of parallel downloads to run.
-            callback (Callable[[Tuple[str, str], Status], None]): Optional allback function to call after each download.
+            callback (Callable[[tuple[str, str], Status], None]): Optional allback function to call after each download.
                          The function must take two arguments: the download tuple and the status of the download thread.
         """
 
@@ -164,9 +167,9 @@ class FileDownloader:
         self.kill_with_fire = False
 
         # Track the download threads
-        future_downloads: Dict[concurrent.futures.Future, Tuple[str, str]] = {}
+        future_downloads: dict[concurrent.futures.Future, tuple[str, str]] = {}
 
-        # List to store *successful* downloads
+        # list to store *successful* downloads
         successful_downloads = []
 
         def successful_download_callback(future: concurrent.futures.Future) -> None:
@@ -252,3 +255,279 @@ class FileDownloader:
                     # Check that we actually downloaded something
                     if not has_content:
                         raise DownloadError(f"Downloaded file '{remote_path}' is empty")
+
+
+##################################################################################################
+#                                                                                                #
+#   Below are helper functions used by the legacy method for extracting pipeline containers      #
+#                                                                                                #
+##################################################################################################
+def rectify_raw_container_matches(raw_findings):
+    """Helper function to rectify the raw extracted container matches into fully qualified container names.
+    If multiple containers are found, any prefixed with http for direct download is prioritized
+
+    Example syntax:
+
+    Early DSL2:
+
+    .. code-block:: groovy
+
+        if (workflow.containerEngine == 'singularity' && !params.singularity_pull_docker_container) {
+            container "https://depot.galaxyproject.org/singularity/fastqc:0.11.9--0"
+        } else {
+            container "quay.io/biocontainers/fastqc:0.11.9--0"
+        }
+
+    Later DSL2:
+
+    .. code-block:: groovy
+
+        container "${ workflow.containerEngine == 'singularity' && !task.ext.singularity_pull_docker_container ?
+            'https://depot.galaxyproject.org/singularity/fastqc:0.11.9--0' :
+            'biocontainers/fastqc:0.11.9--0' }"
+
+    Later DSL2, variable is being used:
+
+    .. code-block:: groovy
+
+        container "${ workflow.containerEngine == 'singularity' && !task.ext.singularity_pull_docker_container ?
+            "https://depot.galaxyproject.org/singularity/${container_id}" :
+            "quay.io/biocontainers/${container_id}" }"
+
+        container_id = 'mulled-v2-1fa26d1ce03c295fe2fdcf85831a92fbcbd7e8c2:afaaa4c6f5b308b4b6aa2dd8e99e1466b2a6b0cd-0'
+
+    DSL1 / Special case DSL2:
+
+    .. code-block:: groovy
+
+        container "nfcore/cellranger:6.0.2"
+
+    """
+    cleaned_matches = []
+
+    # Thanks Stack Overflow for the regex: https://stackoverflow.com/a/3809435/713980
+    url_regex = r"https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)"
+    oras_regex = r"oras:\/\/[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)"
+    # Thanks Stack Overflow for the regex: https://stackoverflow.com/a/39672069/713980
+    docker_regex = r"^(?:(?=[^:\/]{1,253})(?!-)[a-zA-Z0-9-]{1,63}(?<!-)(?:\.(?!-)[a-zA-Z0-9-]{1,63}(?<!-))*(?::[0-9]{1,5})?/)?((?![._-])(?:[a-z0-9._-]*)(?<![._-])(?:/(?![._-])[a-z0-9._-]*(?<![._-]))*)(?::(?![.-])[a-zA-Z0-9_.-]{1,128})?$"
+
+    # at this point, we don't have to distinguish anymore, because we will later prioritize direct downloads over Docker URIs.
+    either_url_or_docker = re.compile(f"{url_regex}|{oras_regex}|{docker_regex}", re.S)
+
+    for _, container_value, search_space, file_path in raw_findings:
+        """
+        Now we need to isolate all container paths (typically quoted strings) from the raw container_value
+
+        For example from:
+
+        "${ workflow.containerEngine == \'singularity\' && !task.ext.singularity_pull_docker_container ?
+        \'https://depot.galaxyproject.org/singularity/ubuntu:20.04\' :
+        \'nf-core/ubuntu:20.04\' }"
+
+        we want to extract
+
+        'singularity'
+        'https://depot.galaxyproject.org/singularity/ubuntu:20.04'
+        'nf-core/ubuntu:20.04'
+
+        The problem is, that we find almost arbitrary notations in container_value,
+        such as single quotes, double quotes and escaped single quotes
+
+        Sometimes target strings are wrapped into a variable to be evaluated by Nextflow,
+        like in the example above.
+        Sometimes the whole string is a variable "${container_name}", that is
+        denoted elsewhere.
+
+        Sometimes the string contains variables which may be defined elsewhere or
+        not even known when downloading, like ${params.genome}.
+
+        "https://depot.galaxyproject.org/singularity/ubuntu:${version}" or
+        "nfcore/sarekvep:dev.${params.genome}"
+
+        Mostly, it is a nested DSL2 string, but it may also just be a plain string.
+
+
+        First, check if container_value is a plain container URI like in DSL1 pipelines
+        or a plain URL like in the old DSL2 convention
+
+        """
+        direct_match = re.match(either_url_or_docker, container_value.strip())
+        if direct_match:
+            cleaned_matches.append(direct_match.group(0))
+            continue  # oh yes, that was plain sailing
+
+        """
+        no plain string, we likely need to break it up further
+
+        [^\"\'] makes sure that the outermost quote character is not matched.
+        (?P<quote>(?<![\\])[\'\"]) again captures the quote character, but not if it is preceded by an escape sequence (?<![\\])
+        (?P<param>(?:.(?!(?<![\\])\1))*.?)\1 is basically what I used above, but again has the (?<![\\]) inserted before \1 to account for escapes.
+        """
+
+        container_value_defs = re.findall(
+            r"[^\"\'](?P<quote>(?<![\\])[\'\"])(?P<param>(?:.(?!(?<![\\])\1))*.?)\1",
+            container_value,
+        )
+
+        """
+        eliminate known false positives and create plain list out of the tuples returned by the regex above
+        example result:
+        ['https://depot.galaxyproject.org/singularity/scanpy:1.7.2--pyhdfd78af_0', 'biocontainers/scanpy:1.7.2--pyhdfd78af_0']
+        """
+        container_value_defs = [
+            capture for _, capture in container_value_defs[:] if capture not in ["singularity", "apptainer"]
+        ]
+
+        """
+        For later DSL2 syntax, container_value_defs should contain both, the download URL and the Docker URI.
+        For earlier DSL2, both end up in different raw_findings, so a subsequent deduplication is necessary anyway
+        to not pull a container that is already downloaded directly.
+
+        At this point, we just add everything that is either a URL or a Docker URI to cleaned matches.
+        """
+
+        valid_containers = list(filter(either_url_or_docker.match, container_value_defs))
+
+        if valid_containers:
+            cleaned_matches = cleaned_matches + valid_containers
+            # Yeah, we have successfully extracted something from this raw_finding, so move on.
+            continue
+
+        """
+        Neither a plain Docker URI nor a DSL2-like definition was found. This is a tricky case, then.
+
+        Some modules declare the container as separate variable. This entails that " instead of ' is used,
+        so container_value will not contain it.
+
+        Therefore, we need to repeat the search over the raw contents, extract the variable name, and use it inside a new regex.
+        This is why the raw search_space is still needed at this level.
+
+        To get the variable name ( ${container_id} in above example ), we match the literal word "container" and use lookbehind (reset the match).
+        Then we skip [^${}]+ everything that is not $ or curly braces. The next capture group is
+        ${ followed by any characters that are not curly braces [^{}]+ and ended by a closing curly brace (}),
+        but only if it's not followed by any other curly braces (?![^{]*}). The latter ensures we capture the innermost
+        variable name.
+        """
+
+        container_definition = re.search(r"(?<=container)[^\${}]+\${([^{}]+)}(?![^{]*})", str(search_space))
+
+        if bool(container_definition) and bool(container_definition.group(1)):
+            pattern = re.escape(container_definition.group(1))
+            # extract the quoted string(s) following the variable assignment
+            container_names = re.findall(rf"{pattern}\s*=\s*[\"\']([^\"\']+)[\"\']", search_space)
+
+            if bool(container_names):
+                if isinstance(container_names, str):
+                    cleaned_matches.append(f"https://depot.galaxyproject.org/singularity/{container_names}")
+
+                elif isinstance(container_names, list):
+                    # this deliberately appends container_names[-1] twice to cleaned_matches
+                    # but deduplication is performed anyway and just setting this_container
+                    # here as well allows for an easy check to see if parsing succeeded.
+                    for container_name in container_names:
+                        cleaned_matches.append(f"https://depot.galaxyproject.org/singularity/{container_name}")
+
+                continue
+
+        # all implemented options exhausted. Nothing left to be done:
+        log.error(
+            f"[red]Cannot parse container string in '{file_path}':\n\n{textwrap.indent(container_value, '    ')}\n\n:warning: Skipping this singularity image."
+        )
+
+    """
+    Loop has finished, now we need to remove duplicates and prioritize direct downloads over containers pulled from the registries
+    """
+    return prioritize_direct_download(cleaned_matches)
+
+
+def prioritize_direct_download(container_list: list[str]) -> list[str]:
+    """
+    Helper function that takes a list of container images (URLs and Docker URIs),
+    eliminates all Docker URIs for which also a URL is contained and returns the
+    cleaned and also deduplicated list.
+
+    Conceptually, this works like so:
+
+    Everything after the last Slash should be identical, e.g. "scanpy:1.7.2--pyhdfd78af_0" in
+    ['https://depot.galaxyproject.org/singularity/scanpy:1.7.2--pyhdfd78af_0', 'biocontainers/scanpy:1.7.2--pyhdfd78af_0']
+
+
+    re.sub('.*/(.*)','\\1',c) will drop everything up to the last slash from c (container_id)
+
+    d.get(k:=re.sub('.*/(.*)','\\1',c),'') assigns the truncated string to k (key) and gets the
+    corresponding value from the dict if present or else defaults to "".
+
+    If the regex pattern matches, the original container_id will be assigned to the dict with the k key.
+    r"^$|(?!^http)" matches an empty string (we didn't have it in the dict yet and want to keep it in either case) or
+    any string that does not start with http. Because if our current dict value already starts with http,
+    we want to keep it and not replace with with whatever we have now (which might be the Docker URI).
+
+    A regex that matches http, r"^$|^http" could thus be used to prioritize the Docker URIs over http Downloads
+
+    We also need to handle a special case: The https:// Singularity downloads from Seqera Containers all end in 'data', although
+    they are not equivalent, e.g.:
+
+    'https://community-cr-prod.seqera.io/docker/registry/v2/blobs/sha256/63/6397750e9730a3fbcc5b4c43f14bd141c64c723fd7dad80e47921a68a7c3cd21/data'
+    'https://community-cr-prod.seqera.io/docker/registry/v2/blobs/sha256/c2/c262fc09eca59edb5a724080eeceb00fb06396f510aefb229c2d2c6897e63975/data'
+
+    Lastly, we want to remove at least a few Docker URIs for those modules, that have an oras:// download link.
+    """
+    d: dict[str, str] = {}
+    seqera_containers_http: list[str] = []
+    seqera_containers_oras: list[str] = []
+    all_others: list[str] = []
+
+    for c in container_list:
+        if bool(re.search(r"/data$", c)):
+            seqera_containers_http.append(c)
+        elif bool(re.search(r"^oras://", c)):
+            seqera_containers_oras.append(c)
+        else:
+            all_others.append(c)
+
+    for c in all_others:
+        if re.match(r"^$|(?!^http)", d.get(k := re.sub(".*/(.*)", "\\1", c), "")):
+            log.debug(f"{c} matches and will be saved as {k}")
+            d[k] = c
+
+    combined_with_oras = reconcile_seqera_container_uris(seqera_containers_oras, list(d.values()))
+
+    # combine deduplicated others (Seqera containers oras, http others and Docker URI others) and Seqera containers http
+    return sorted(list(set(combined_with_oras + seqera_containers_http)))
+
+
+def reconcile_seqera_container_uris(prioritized_container_list: list[str], other_list: list[str]) -> list[str]:
+    """
+    Helper function that takes a list of Seqera container URIs,
+    extracts the software string and builds a regex from them to filter out
+    similar containers from the second container list.
+
+    prioritzed_container_list = [
+    ...     "oras://community.wave.seqera.io/library/multiqc:1.25.1--f0e743d16869c0bf",
+    ...     "oras://community.wave.seqera.io/library/multiqc_pip_multiqc-plugins:e1f4877f1515d03c"
+    ... ]
+
+    will be cleaned to
+
+    ['library/multiqc:1.25.1', 'library/multiqc_pip_multiqc-plugins']
+
+    Subsequently, build a regex from those and filter out matching duplicates in other_list:
+    """
+    if not prioritized_container_list:
+        return other_list
+    else:
+        # trim the URIs to the stem that contains the tool string, assign with Walrus operator to account for non-matching patterns
+        trimmed_priority_list = [
+            match.group()
+            for c in set(prioritized_container_list)
+            if (match := (re.search(r"library/.*?:[\d.]+", c) if "--" in c else re.search(r"library/[^\s:]+", c)))
+        ]
+
+        # build regex
+        prioritized_containers = re.compile("|".join(f"{re.escape(c)}" for c in trimmed_priority_list))
+
+        # filter out matches in other list
+        filtered_containers = [c for c in other_list if not re.search(prioritized_containers, c)]
+
+        # combine prioritized and regular container lists
+        return sorted(list(set(prioritized_container_list + filtered_containers)))
